Cleanup of README.md (#342)
* Reformatted README.md

This is an initial cleanup attempt of the README.md file.  All
non-code lines have been wrapped to an 80 column width, and some
markdown was cleaned up.  Further work is needed.

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Added 'bash' info string to appropriate fenced code blocks.

Numerous fenced code blocks did not have a
[info string](https://github.github.com/gfm/#info-string), which
meant that the enclosed code block wouldn't be rendered correctly.
As it turned out, all of these were intended to be run in a bash
shell, so I added that as their info strings.

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* WIP: Updated the scripts in the README.md file to use environment variables.

The scripts weren't directly executable before, requiring manual
substitution of various 'variables'.  This was unfortunate as Github
makes copying and pasting code blocks simple and straightforward.

The changes in this commit make it possible to copy/paste each example
into it's own shell, which should make it slightly easier for new users
to start using the bridge.

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* fix: Unwrapped lines per request from @gbiggs.

@gbiggs commented at https://github.com/ros2/ros1_bridge/pull/342#issuecomment-1029508012
that files in ROS2 are intentionally not wrapped to make handling
diffs in PRs easier.  I've copied in the relevant original lines
to unwrap the lines I wrapped earlier.

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Accepted suggestion from @gbiggs

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* Update README.md

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* The service name might provide more debug information than host:port details (#346)

* The service name might provide more debug information than host:port details

Signed-off-by: Tim Clephas <tim.clephas@nobleo.nl>
Co-authored-by: Tomoya Fujita <Tomoya.Fujita@sony.com>
Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

* fix: Changed 'galactic' to 'rolling'

Changes due to suggestions from @gbiggs in the discussion of [PR 342](https://github.com/ros2/ros1_bridge/pull/342#discussion_r834287510)

Signed-off-by: Cem Karan <cem.f.karan.civ@army.mil>

Co-authored-by: Geoffrey Biggs <gbiggs@killbots.net>
Co-authored-by: Tim Clephas <tim.clephas@nobleo.nl>
Co-authored-by: Tomoya Fujita <Tomoya.Fujita@sony.com>